# swagger/v1/swagger.yml
swagger: "2.0"
info:
  description: "家事の管理ツール：ChoreiのAPIエンドポイント"
  version: "1.0.0"
  title: "Chorei API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: ""
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "3.113.114.91"
basePath: "/api/v1"
...
securityDefinitions:
  access-token:
    type: "apiKey"
    name: "access-token"
    in: "header"
  uid:
    type: "apiKey"
    name: "uid"
    in: "header"
  client:
    type: "apiKey"
    name: "client"
    in: "header"
schemes:
- "http"
paths:
  /auth:
    post:
      tags:
      - "user"
      summary: "SignUp用"
      description: "name, email, passwordでuserのsignupする"
      operationId: "auth"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "sign_up"
        description: "サインアップ用パラメータ"
        required: true
        schema:
          type: object
          properties:
            name:
              type: "string"
              example: "test user"
            email:
              type: "string"
              example: "test@example.com"
            password:
              type: "password"
              format: "password"
              example: "password"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "success"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  provider:
                    type: "string"
                    example: "email"
                  uid: 
                    type: "string"
                    example: "fource@test.com"
                  allow_password_change:
                    type: "boolean"
                    example: false
                  name: 
                    type: "string"
                    example: "fource"
                  image: 
                    type: "object"
                    properties:
                      url: 
                        type: "string"
                        exampmle: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  email:
                    type: "string"
                    example: "fource@test.com"
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        422:
          description: "失敗時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "error"
              data:
                type: "object"
                properties:
                  some-properties: "some properties..."
              errors:
                type: "object"
                properties:
                  full_messages:
                    type: "array"
                    items:
                      type: "string"
                      example: ["password can't be blank","Email has already been taken"]
    delete:
      tags:
      - "user"
      summary: "ユーザー削除用"
      description: "destroy user"
      operationId: "auth/registrations#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      parameters:
      - in: "header"
        name: "access-token"
        description: "access-token"
        required: true
        type: "string"
      - in: "header"
        name: "uid"
        description: "uid(email)"
        required: true
        type: "string"
      - in: "header"
        name: "client"
        description: "client"
        required: true
        type: "string"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: true
        404:
          description: "トークンは不正だった場合のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: false
              errors:
                type: "array"
                items:
                  type: "string"
                  example: "User was not found or was not logged in."
...
  /auth/sign_out:
    delete:
      tags:
      - "user"
      summary: "logout用"
      description: "logout用"
      operationId: "devise_token_auth/sessions#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "log out"
        description: "log out"
        required: true
        schema:
          type: object
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: true
        404:
          description: "ログアウト済なのにログアウトした場合のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: false
              errors:
                type: "array"
                items:
                  type: "string"
                  example: "User was not found or was not logged in."
      security:
      - access-token: 
        uid:
        client:
...
  /members/{member_id}:
    delete:
      tags:
      - "member"
      summary: "members#destroy"
      description: "pathでidを指定してmemberを削除"
      operationId: "Members#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "member_id"
        description: "member#show member_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Member"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  group_id:
                    type: "integer"
                    example: 1
                  user_id:
                    type: "integer"
                    example: 10
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(実在しないmember_id)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Member is Not Found"
  /tags:
    post:
      tags:
      - "tag"
      summary: "tags#create"
      description: "paramsでnameを指定してtagをcreate"
      operationId: "tags#create"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "tag_create"
        description: "tags#create"
        required: true
        schema:
          type: object
          properties:
            name:
              type: "string"
              example: "test user"
      responses:
        201:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 201
              message:
                type: "string"
                example: "Created: Tag"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  name:
                    type: "string"
                    example: "test tag"
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        400:
          description: "失敗時のレスポンス(nameが空欄)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 400
              message:
                type: "string"
                example: "Nameを入力してください"
      security:
      - access-token: 
        uid:
        client:
    get:
      tags:
      - "tag"
      summary: "tags#index"
      description: "tagの一覧を表示"
      operationId: "tags#index"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Tag"
              data:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    name:
                      type: "string"
                      example: "test tag"
                    created_at:
                      type: "string"
                      example: "2019-06-01T12:01:22.868Z"
                    updated_at:
                      type: "string"
                      example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(tagが存在しない)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Tag is Not Found"
      security:
      - access-token: 
        uid:
        client:
    delete:
      tags:
      - "tag"
      summary: "tags#destroy"
      description: "pathで指定したIDのTagを削除"
      operationId: "tags#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Tag"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  name:
                    type: "string"
                    example: "test tag"
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(tagが存在しない)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Tag is Not Found"
      security:
      - access-token: 
        uid:
        client:
...
  /groups/{group_id}:
    put:
      tags:
      - "group"
      summary: "Group#Update"
      description: "Pathでidを指定してUpdate"
      operationId: "groups#update"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "group#show group_id"
        required: true
      - in: "body"
        name: "group#update"
        description: "group#update parameters"
        required: true
        schema:
          type: object
          properties:
            name:
              type: "string"
              example: "test group"
            image:
              type: "string"
              example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Group"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  name: 
                    type: "string"
                    example: "updated test group"
                  image: 
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  owner_id: 
                    type: "integer"
                    example: 2
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        400:
          description: "失敗時のレスポンス(実在しないowner_id, 名前の長さ(1~100文字外)、画像がbase64形式でない場合)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 400
              message:
                type: "string"
                example: "Bad Request"
              data:
                type: "array"
                example: [ 名称を入力してください, 画像のMIMEタイプが不正です, オーナーを入力してください]
        404:
          description: "失敗時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Group is Not Found"
      security:
      - access-token: 
        uid:
        client:
    get:
      tags:
      - "group"
      summary: "Group#Show"
      description: "Pathでidを指定してGroupを取得"
      operationId: "groups#show"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "group#show group_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Group"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  owner_id: 
                    type: "integer"
                    example: 2
                  name: 
                    type: "string"
                    example: "test group"
                  image: 
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Group is Not Found"
      security:
      - access-token: 
        uid:
        client:
    delete:
      tags:
      - "group"
      summary: "Group#Destroy"
      description: "Pathでidを指定してGroupを削除"
      operationId: "groups#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "group#destroy group_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Group"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  owner_id: 
                    type: "integer"
                    example: 2
                  name: 
                    type: "string"
                    example: "test group"
                  image: 
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Group is Not Found"
      security:
      - access-token: 
        uid:
        client:
...
  /assigns/{assign_id}:
    delete:
      tags:
      - "assign"
      summary: "assigns#destroy"
      description: "pathでassign_idを指定して削除"
      operationId: "assigns#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "assign_id"
        description: "assign_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Assign"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  member_id:
                    type: "integer"
                    example: 2
                  job_id:
                    type: "integer"
                    example: 8
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(実在しないAssignを指定)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Assign is Not Found  "
        406:
          description: "失敗時のレスポンス(当該assignされているjobが属しているgroupにcurrent_userが属していないとき)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 406
              message:
                type: "string"
                example: "Assign is Not Acceptable"
      security:
      - access-token: 
        uid:
        client:
...
  /taggings/{tagging_id}:
    delete:
      tags:
      - "tagging"
      summary: "taggings#destroy"
      description: "pathでidを指定してtaggingをdestroyする"
      operationId: "taggings#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "tagging_id"
        description: "tagging_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Tagging"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  job_id:
                    type: "string"
                    example: "weekly"
                  tag_id:
                    type: integer
                    example: 1
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(実在しないTaggingを指定した時)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Tagging is Not Found"
        406:
          description: "失敗時のレスポンス(current_userがjobのmemberでない)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 406
              message:
                type: "string"
                example: "Member is Not Acceptable"
      security:
      - access-token: 
        uid:
        client:
...
  /stocks/{stock_id}:
    delete:
      tags:
      - "stock"
      summary: "stocks#destroy"
      description: "pathでstock_idを指定"
      operationId: "stocks#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "stock_id"
        description: "stock_id"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success: Stock"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  job_id:
                    type: "string"
                    example: "weekly"
                  user_id:
                    type: integer
                    example: 1
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "失敗時のレスポンス(実在しないStockを指定した時)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 404
              message:
                type: "string"
                example: "Stock is Not Found"
        406:
          description: "失敗時のresponse(current_user以外のstockを指定したとき)"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: 406
              message:
                type: "string"
                example: "Member is Not Acceptable"
      security:
      - access-token: 
        uid:
        client: